fix!: updated cookie expiry date
This was a really nasty bug. Basically, the session cookie was set to
expire after 1 hour. But if the server is in a different time zone than
the client, the cookie expires before it's used. This patch makes
cookies expire after 10 days, making sure everyone can log in and use
the app.